Extract SDK for 3rd-party use.
Barely functional for CSUN13 demo
